feat: Add `Endpoint::connect_to_any`
This implements the 'bootstrap' process of contacting a set of nodes and
returning the first one to which we successfully connect.
`QuicP2p::bootstrap` is now implemented in terms of this function.
